@@ -1,20 +1,27 @@
 {
-  "name": "short-link",
+  "name": "short-lnk",
   "private": true,
   "scripts": {
-    "start": "meteor run"
+    "start": "meteor run",
+    "test": "meteor test --driver-package=practicalmeteor:mocha"
   },
   "dependencies": {
     "babel-runtime": "6.18.0",
     "meteor-node-stubs": "~0.2.0",
     "moment": "^2.17.1",
-    "prop-types": "^15.5.10",
-    "react": "^15.4.1",
+    "react": "^15.5.4",
+    "react-addons-pure-render-mixin": "^15.4.1",
     "react-dom": "^15.4.1",
     "react-router": "^3.0.0",
+    "react-test-renderer": "^15.5.4",
     "simpl-schema": "0.0.3"
   },
   "engines": {
     "node": "4.6.2"
+  },
+  "devDependencies": {
+    "enzyme": "^2.7.1",
+    "expect": "^1.20.2",
+    "react-addons-test-utils": "^15.4.1"
   }
 }
