@@ -1,20 +1,25 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import { Accounts } from 'meteor/accounts-base';
+import { createContainer } from 'meteor/react-meteor-data';
 
-const PrivateHeader = (props) => {
-    return(
-      <div className="header">
-        <div className="header__content">
-          <h1 className="header__title">{props.title}</h1>
-          <button className="button button--link-text" onClick={() => Accounts.logout()}>Logout</button>
+export const PrivateHeader = (props) => {
+  return (
+    <div className="header">
+      <div className="header__content">
+        <h1 className="header__title">{props.title}</h1>
+        <button className="button button--link-text" onClick={() => props.handleLogout()}>Logout</button>
       </div>
-      </div>
-    );
+    </div>
+  );
 };
 
 PrivateHeader.propTypes = {
-  title: React.PropTypes.string.isRequired
+  title: React.PropTypes.string.isRequired,
+  handleLogout: React.PropTypes.func.isRequired
 };
 
-export default PrivateHeader;
+export default createContainer(() => {
+  return {
+    handleLogout: () => Accounts.logout()
+  };
+}, PrivateHeader);
