@@ -1,8 +1,9 @@
 import React from 'react';
 import { Link } from 'react-router';
 import { Meteor } from 'meteor/meteor';
+import { createContainer } from 'meteor/react-meteor-data';
 
-export default class Login extends React.Component {
+export class Login extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -15,9 +16,9 @@ export default class Login extends React.Component {
     let email = this.refs.email.value.trim();
     let password = this.refs.password.value.trim();
 
-    Meteor.loginWithPassword({email}, password, (err) => {
+    this.props.loginWithPassword({email}, password, (err) => {
       if (err) {
-        this.setState({error: 'Unable to login. Check email and password'});
+        this.setState({error: 'Unable to login. Check email and password.'});
       } else {
         this.setState({error: ''});
       }
@@ -37,9 +38,19 @@ export default class Login extends React.Component {
             <button className="button">Login</button>
           </form>
 
-            <Link to="/signup">Need an account bro?</Link>
+          <Link to="/signup">Need an account?</Link>
         </div>
       </div>
     );
   }
 }
+
+Login.propTypes = {
+  loginWithPassword: React.PropTypes.func.isRequired
+};
+
+export default createContainer(() => {
+  return {
+    loginWithPassword: Meteor.loginWithPassword
+  };
+}, Login);
