@@ -2,11 +2,11 @@ import React from 'react';
 import { Link } from 'react-router';
 
 export default () => {
-  return(
+  return (
     <div className="boxed-view">
       <div className="boxed-view__box">
-        <h1>404 Page Not Found</h1>
-        <p>Hmmm, we're unable to find that page.</p>
+        <h1>404 - Page Not Found</h1>
+        <p>We're unable to find that page.</p>
         <Link to="/" className="button button--link">HEAD HOME</Link>
       </div>
     </div>
